move HandleErrorOpts to fs/types/abstract
